+++ mountpoint-s3/tests/cli.rs
@@ -143,18 +143,6 @@ fn s3_uri_as_bucket_name() -> Result<(), Box<dyn std::error::Error>> {
 }
 
 #[test]
-fn invalid_profile() -> Result<(), Box<dyn std::error::Error>> {
-    let dir = assert_fs::TempDir::new()?;
-    let mut cmd = Command::cargo_bin("mount-s3")?;
-
-    cmd.arg("test-bucket").arg(dir.path()).arg("--profile").arg("INVALID");
-    let error_message = "invalid AWS credentials";
-    cmd.assert().failure().stderr(predicate::str::contains(error_message));
-
-    Ok(())
-}
-
-#[test]
 fn validate_log_files_permissions() -> Result<(), Box<dyn std::error::Error>> {
     let mut cmd = Command::cargo_bin("mount-s3")?;
     let dir = assert_fs::TempDir::new()?;